+++ music/core/filters/__init__.py
@@ -2,6 +2,8 @@
 
 from .adsr import adsr, adsr_stereo, adsr_vibrato
 from .fade import cross_fade, fade
+from .design import (band_pass, band_reject, fraction_of, high_pass,
+                     low_pass)
 from .impulse_response import fir, iir
 from .localization import localize, localize2, localize_linear
 from .loud import loud, louds
@@ -9,6 +11,11 @@ from .reverb import reverb
 from .stretches import stretches
 
 __all__ = [
+    'band_pass',
+    'band_reject',
+    'fraction_of',
+    'high_pass',
+    'low_pass',
     'adsr',
     'adsr_stereo',
     'adsr_vibrato',